move: 파일명, 컴포넌트명 변경 및 폴더 이동
