+++ src/app/(public)/mate/_components/select/filterDateSelect.tsx
@@ -1,0 +1,26 @@
+import { Input } from "@nextui-org/react";
+
+interface FilterDateSelectProps {
+  label: string;
+  selected: string | undefined;
+  onSelect: (value: string) => void;
+}
+
+const FilterDateSelect = ({ label, onSelect, selected }: FilterDateSelectProps) => {
+  const handleChange = (e: React.ChangeEvent<HTMLInputElement>) => {
+    onSelect(e.target.value);
+  };
+
+  return (
+    <Input
+      type="date"
+      label={label}
+      value={selected || ""}
+      radius="full"
+      onChange={handleChange}
+      className="mt-5 w-full"
+    />
+  );
+};
+
+export default FilterDateSelect;
