+++ src/app/(public)/mate/_components/select/filterWeightSelect.tsx
@@ -1,16 +1,15 @@
 // import { Slider } from "@nextui-org/react";
 import { useState } from "react";
-import {Slider, SliderValue} from "@nextui-org/react";
+import { Slider, SliderValue } from "@nextui-org/react";
 
-interface FilterWeightChipProps {
+interface FilterWeightSelectProps {
   label: string;
   selected: string | null;
   onSelect: (value: string) => void;
 }
 
-const FilterWeightChip = ({ label, selected, onSelect }: FilterWeightChipProps) => {
+const FilterWeightSelect = ({ label, selected, onSelect }: FilterWeightSelectProps) => {
   const selectedValue = selected ? parseInt(selected, 10) : 0;
-
 
   const handleChange = (value: number | number[]) => {
     // value가 배열이 아닌 경우에만 처리
@@ -29,7 +28,7 @@ const FilterWeightChip = ({ label, selected, onSelect }: FilterWeightChipProps) 
         defaultValue={0}
         value={selectedValue}
         label={label}
-        aria-label={label}        
+        aria-label={label}
         onChange={handleChange}
         getValue={(weight) => `${weight}kg 이상`}
         showTooltip={true}
@@ -38,27 +37,12 @@ const FilterWeightChip = ({ label, selected, onSelect }: FilterWeightChipProps) 
           filler: "bg-mainColor",
           label: "text-[1rem] font-[400]",
           value: "text-sm  font-[400] text-gray-500",
-          thumb: [
-            "bg-mainColor",
-            "border border-mainColor",
-            "shadow-md",
-          ],
-          track: "bg-gray-200",
+          thumb: ["bg-mainColor", "border border-mainColor", "shadow-md"],
+          track: "bg-gray-200"
         }}
       />
-      {/* <Slider 
-        // label={label}
-        color="foreground"
-        step={10}
-        maxValue={40}
-        minValue={0}
-        value={value} 
-        onChange={handleChange}
-        className="max-w-md"
-        getValue={(value) => `${value}kg 이상`}
-      /> */}
     </div>
   );
 };
 
-export default FilterWeightChip;
+export default FilterWeightSelect;
